move `JavaScriptFilePredicate` back
