+++ sonar-plugin/sonar-javascript-plugin/src/test/java/org/sonar/plugins/javascript/JavaScriptFilePredicateTest.java
@@ -17,16 +17,18 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.plugins.javascript.bridge;
+package org.sonar.plugins.javascript;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.sonar.plugins.javascript.bridge.JavaScriptFilePredicate.isTypeScriptFile;
+import static org.sonar.plugins.javascript.JavaScriptFilePredicate.isTypeScriptFile;
 
 import java.io.File;
+import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 import org.sonar.api.batch.fs.FilePredicate;
@@ -43,7 +45,7 @@ class JavaScriptFilePredicateTest {
   private static final String newLine = System.lineSeparator();
 
   @Test
-  void testJavaScriptPredicate() {
+  void testJavaScriptPredicate() throws IOException {
     DefaultFileSystem fs = new DefaultFileSystem(baseDir);
     fs.add(createInputFile(baseDir, "a.js"));
     fs.add(createInputFile(baseDir, "b.ts"));
@@ -123,13 +125,13 @@ class JavaScriptFilePredicateTest {
       .stream()
       .filter(f -> !isTypeScriptFile(f))
       .map(InputFile::filename)
-      .toList();
+      .collect(Collectors.toList());
     assertThat(filenames)
       .containsExactlyInAnyOrder("a.js", "c.vue", "d.vue", "e.vue", "f.vue", "h.vue", "j.jsx");
   }
 
   @Test
-  void testTypeScriptPredicate() {
+  void testTypeScriptPredicate() throws IOException {
     DefaultFileSystem fs = new DefaultFileSystem(baseDir);
     fs.add(createInputFile(baseDir, "a.js"));
     fs.add(createInputFile(baseDir, "b.ts"));
@@ -200,12 +202,12 @@ class JavaScriptFilePredicateTest {
       .stream()
       .filter(JavaScriptFilePredicate::isTypeScriptFile)
       .map(InputFile::filename)
-      .toList();
+      .collect(Collectors.toList());
     assertThat(filenames).containsExactlyInAnyOrder("b.ts", "e.vue", "f.vue", "j.tsx");
   }
 
   @Test
-  void testYamlPredicate() {
+  void testYamlPredicate() throws IOException {
     var baseYamlFile =
       "".concat("apiVersion: apps/v1")
         .concat(newLine)
@@ -235,7 +237,7 @@ class JavaScriptFilePredicateTest {
     List<File> files = new ArrayList<>();
     fs.files(predicate).forEach(files::add);
 
-    var filenames = files.stream().map(File::getName).toList();
+    List<String> filenames = files.stream().map(File::getName).collect(Collectors.toList());
     assertThat(filenames)
       .containsExactlyInAnyOrder(
         "single-quote.yaml",
@@ -249,22 +251,22 @@ class JavaScriptFilePredicateTest {
   void testIsTypeScriptFile() {
     var tsFile = TestInputFileBuilder
       .create("", "file.ts")
-      .setLanguage("ts")
+      .setLanguage(TypeScriptLanguage.KEY)
       .build();
     assertThat(isTypeScriptFile(tsFile)).isTrue();
     var jsFile = TestInputFileBuilder
       .create("", "file.js")
-      .setLanguage("js")
+      .setLanguage(JavaScriptLanguage.KEY)
       .build();
     assertThat(isTypeScriptFile(jsFile)).isFalse();
     var vueFile = TestInputFileBuilder
       .create("", "file.vue")
-      .setLanguage("js")
+      .setLanguage(JavaScriptLanguage.KEY)
       .build();
     assertThat(isTypeScriptFile(vueFile)).isFalse();
     var tsVueFile = TestInputFileBuilder
       .create("", "file.vue")
-      .setLanguage("js")
+      .setLanguage(JavaScriptLanguage.KEY)
       .setContents("<script lang='ts'>")
       .build();
     assertThat(isTypeScriptFile(tsVueFile)).isTrue();
@@ -275,11 +277,11 @@ class JavaScriptFilePredicateTest {
     var fs = new DefaultFileSystem(baseDir);
     var tsFile = TestInputFileBuilder
       .create("", "file.ts")
-      .setLanguage("ts")
+      .setLanguage(TypeScriptLanguage.KEY)
       .build();
     var jsFile = TestInputFileBuilder
       .create("", "file.js")
-      .setLanguage("js")
+      .setLanguage(JavaScriptLanguage.KEY)
       .build();
     var f = TestInputFileBuilder.create("", "file.cpp").setLanguage("c").build();
     var predicate = JavaScriptFilePredicate.getJsTsPredicate(fs);
@@ -293,21 +295,22 @@ class JavaScriptFilePredicateTest {
   }
 
   private static InputFile createInputFile(Path baseDir, String relativePath, String content) {
-    return new TestInputFileBuilder("moduleKey", relativePath)
+    InputFile file = new TestInputFileBuilder("moduleKey", relativePath)
       .setModuleBaseDir(baseDir)
       .setType(Type.MAIN)
       .setLanguage(getLanguage(relativePath))
       .setCharset(StandardCharsets.UTF_8)
       .setContents(content)
       .build();
+    return file;
   }
 
   private static String getLanguage(String path) {
     String fileExtension = path.substring(path.indexOf("."));
-    if (".js,.jsx,.cjs,.mjs,.vue".contains(fileExtension)) {
-      return "js";
-    } else if (".ts,.tsx,.cts,.mts".contains(fileExtension)) {
-      return "ts";
+    if (JavaScriptLanguage.FILE_SUFFIXES_DEFVALUE.contains(fileExtension)) {
+      return JavaScriptLanguage.KEY;
+    } else if (TypeScriptLanguage.FILE_SUFFIXES_DEFVALUE.contains(fileExtension)) {
+      return TypeScriptLanguage.KEY;
     } else {
       return path.split("\\.")[1];
     }
